refactor: remove util class
 * Move helper methods to needed test classes
 * Remove unnused methods
 * Remove utility at the end
